Simplified Chinese and German translations

+++ UserInterface/SettingsUIBuilder.cs
@@ -155,19 +155,21 @@ namespace HideItBobby.UserInterface
                 {
                     try
                     {
+                        var dePath = Path.Combine(Paths.TranslationsDir, "hide_it_bobby.de.xml");
+                        if (File.Exists(dePath)) File.Delete(dePath);
+                        File.WriteAllText(dePath, Properties.Resources.hide_it_bobby_de, System.Text.Encoding.UTF8);
+
                         var enPath = Path.Combine(Paths.TranslationsDir, "hide_it_bobby.en.xml");
-                        if (!File.Exists(enPath))
-                        {
-                            File.Delete(enPath);
-                            File.WriteAllText(enPath, Properties.Resources.hide_it_bobby_en, System.Text.Encoding.UTF8);
-                        }
+                        if (File.Exists(enPath)) File.Delete(enPath);
+                        File.WriteAllText(enPath, Properties.Resources.hide_it_bobby_en, System.Text.Encoding.UTF8);
 
                         var plPath = Path.Combine(Paths.TranslationsDir, "hide_it_bobby.pl.xml");
-                        if (!File.Exists(plPath))
-                        {
-                            File.Delete(plPath);
-                            File.WriteAllText(plPath, Properties.Resources.hide_it_bobby_pl, System.Text.Encoding.UTF8);
-                        }
+                        if (File.Exists(plPath)) File.Delete(plPath);
+                        File.WriteAllText(plPath, Properties.Resources.hide_it_bobby_pl, System.Text.Encoding.UTF8);
+
+                        var zhPath = Path.Combine(Paths.TranslationsDir, "hide_it_bobby.zh.xml");
+                        if (File.Exists(zhPath)) File.Delete(zhPath);
+                        File.WriteAllText(zhPath, Properties.Resources.hide_it_bobby_zh, System.Text.Encoding.UTF8);
                     }
                     catch (Exception e)
                     {
